+++ NEWS.md
@@ -13,6 +13,7 @@
 - Data from `getjagsdata()` contains `studyID` and has been added to `mbnma` objects
 - Added studyID column to output from `devplot()` and `fitplot()`
 - `plot.nodesplit()` scales y-axis if density is >50 times larger in panel with highest density than in panel with lowest density. This improves legibility of the graph.
+- All nodesplit models now return object of `class("nodesplit")`
 
 # MBNMAdose 0.2.7